refactor(pomodoro): move pomodoro logics to context
