+++ app/dashboard/pomodoro/PomodoroIndex.tsx
@@ -12,21 +12,20 @@ import React, { useEffect, useState } from "react";
 import CycleVisualizer from "./CycleVisualizer";
 import { formatTime } from "@/src/lib/tusktask/utils/format/formatTIme";
 import TimerControls from "./TimerControls";
-
-// Constants for focus and break times in seconds
-const FOCUS_TIME = 25 * 60; // 25 minutes
-const BREAK_TIME = 5 * 60; // 5 minutes
+import usePomodoroContext from "@/src/lib/tusktask/hooks/context/usePomodoroContext";
 
 // Main Pomodoro timer component
 const PomodoroIndex = () => {
-  const [time, setTime] = useState(FOCUS_TIME); // Initial time set to 25 minutes
-  const [isRunning, setIsRunning] = useState(false);
-  const [cycle, setCycle] = useState<"focus" | "break">("focus");
+  // Pull triggers from Notification context
+  const { stopAllSounds } = useNotificationContext();
 
-  const { triggerToast, stopAllSounds } = useNotificationContext();
+  // Pull setters from tasks context
   const { setNewTaskDialogOpen, setIsPomodoroTask } = useTasksContext();
+
+  // Pull session from auth context
   const { data: session } = useSession();
 
+  // Initialize Query
   const { data } = useQuery({
     queryKey: ["tasks", "pomodoro"],
     queryFn: () => {
@@ -39,67 +38,9 @@ const PomodoroIndex = () => {
     enabled: !!session?.user.id,
   });
 
-  // Timer decrement logic
-  useEffect(() => {
-    let interval: string | number | NodeJS.Timeout | undefined;
-    if (isRunning && time > 0) {
-      interval = setInterval(() => {
-        setTime((prevTime) => {
-          const next = Math.max(prevTime - 1, 0);
-          document.title = `${formatTime(next)} | ${cycle[0].toUpperCase() + cycle.slice(1)}`;
-          return next;
-        });
-      }, 1000);
-    }
-    return () => clearInterval(interval);
-  }, [isRunning, time, cycle]);
-
-  // Switch cycle when time reaches zero
-  useEffect(() => {
-    if (time === 0) {
-      handleCycleSwitch();
-    }
-  }, [time]);
-
-  // Handle switching between focus and break cycles
-  const handleCycleSwitch = () => {
-    const newCycle = cycle === "focus" ? "break" : "focus";
-    setCycle(newCycle);
-    setTime(newCycle === "focus" ? FOCUS_TIME : BREAK_TIME);
-    setIsRunning(false);
-    const title = newCycle === "break" ? "Take A Break" : "Let's Roll Again";
-    const description =
-      newCycle === "break"
-        ? "Time for take a break, recharge your focus!"
-        : "It's time to focus again and finish your task!";
-    triggerToast(
-      {
-        type: "reminder",
-        title: title,
-        description: description,
-      },
-      true,
-      "alarm"
-    );
-    document.title =
-      newCycle === "focus"
-        ? "It's time to Focus again!"
-        : "Good job, now take a break!";
-    new Notification(title, {
-      body: description,
-    });
-  };
-
-  // Reset timer to initial value of current cycle
-  const resetTimer = () => {
-    setIsRunning(false);
-    setTime(cycle === "focus" ? FOCUS_TIME : BREAK_TIME);
-  };
-
-  // Skip to next cycle
-  const skipCycle = () => {
-    handleCycleSwitch();
-  };
+  // Pull controllers and states from PomodoroContext
+  const { cycle, isRunning, time, setIsRunning, resetTimer, skipCycle } =
+    usePomodoroContext();
 
   return (
     <div className="space-y-6">
